Finished menu display functionality

+++ src/main/java/org/launchcode/controllers/CheeseController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import javax.validation.Valid;
+import javax.validation.constraints.Null;
 
 /**
  * Created by LaunchCode
@@ -72,10 +73,11 @@ public class CheeseController {
     @RequestMapping(value = "remove", method = RequestMethod.POST)
     public String processRemoveCheeseForm(@RequestParam int[] cheeseIds) {
 
-        for (int cheeseId : cheeseIds) {
-            cheeseDao.delete(cheeseId);
+        if (cheeseIds.length > 0) {
+            for (int cheeseId : cheeseIds) {
+                cheeseDao.delete(cheeseId);
+            }
         }
-
         return "redirect:";
     }
